requirements.txt
ortools~=9.5.2237
shapely~=2.0.1
argparse~=1.4.0
matplotlib~=3.5.1
numpy~=1.21.4
PyYAML~=6.0